Se agrega modificar y eliminar producto.

+++ src/services/productService.js
@@ -1,5 +1,6 @@
 // src/services/productService.js
 const API_URL = process.env.REACT_APP_API_URL || 'http://localhost:4000';
+const getToken = () => localStorage.getItem('userToken');
 
 export const productService = {
     getAll: async (categoria = null) => {
@@ -32,9 +33,13 @@ export const productService = {
     },
 
     update: async (id, productData) => {
+        const token = getToken();
         const response = await fetch(`${API_URL}/productos/${id}`, {
             method: 'PUT',
-            headers: { 'Content-Type': 'application/json' },
+            headers: {
+                'Content-Type': 'application/json',
+                'Authorization': `Bearer ${token}`
+            },
             body: JSON.stringify(productData)
         });
 
@@ -44,10 +49,21 @@ export const productService = {
     },
 
     delete: async (id) => {
+        const token = getToken();
+
+        if (!token) throw new Error("No estás autenticado");
+
         const response = await fetch(`${API_URL}/productos/${id}`, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${token}`
+            }
         });
-        if (!response.ok) throw new Error('Error al eliminar');
+
+        if (!response.ok) {
+            const data = await response.json();
+            throw new Error(data.mensaje || 'Error al eliminar');
+        }
         return true;
     }
 };
